+++ packages/pkg-cli/src/index.ts
@@ -2,7 +2,7 @@ import "@business-as-code/core";
 import "@business-as-code/plugin-core-essentials";
 import "@business-as-code/plugin-core-git";
 // if (process.env.NODE_ENV === 'development') {
-  //   import "@business-as-code/plugin-core-tests";
+  //   import "@business-as-code/plugin-dev-tests";
   //   import "@business-as-code/plugin-dev-changesets";
   //   import "@business-as-code/plugin-dev-essentials";
 // }