+++ packages/tests/pkg-tests-core/src/schematics/schematics-run-command.ts
@@ -85,7 +85,7 @@ export class SchematicsRunCommand extends BaseCommand<
 
 
     const res = await schematicsService.runSchematic({
-      // address: `@business-as-code/plugin-core-tests#namespace=repositories-create`,
+      // address: `@business-as-code/plugin-dev-tests#namespace=repositories-create`,
       address: context.cliOptions.flags.schematicsAddress,
       context: context,
       options: (context.cliOptions.flags as any).schematicOptions,