+++ packages/tests/plugin-dev-tests/src/commands/bac-tests/schematics-run.ts
@@ -154,7 +154,7 @@ export class BacTestsSchematicsRun extends BaseCommand<
     });
 
     const res = await schematicsService.runSchematic({
-      // address: `@business-as-code/plugin-core-tests#namespace=repositories-create`,
+      // address: `@business-as-code/plugin-dev-tests#namespace=repositories-create`,
       address: context.cliOptions.flags.schematicsAddress,
       context,
       options: context.cliOptions.flags["schematicOptions"]!,