+++ packages/pkg-fslib/CHANGELOG.md
@@ -1,5 +1,84 @@
 # @business-as-code/fslib
 
+## 0.0.0-bollards-20230521111204
+
+### Patch Changes
+
+- new message
+- snapshot
+
+## 0.0.0-bollards-20230521110729
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521110022
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521105340
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521104312
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521102912
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521102656
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521102149
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521101719
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521101234
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521101045
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521100737
+
+### Patch Changes
+
+- new message
+
+## 0.0.0-bollards-20230521084807
+
+### Patch Changes
+
+- new message
+
 ## 0.0.0-bollards-20230521084442
 
 ### Patch Changes