+++ packages/tests/pkg-tests-git-mock-server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # git-http-mock-server
 
+## 0.0.0-bollards-20230521111204
+
+### Patch Changes
+
+- snapshot
+
 ## 0.0.0-bollards-20230519070635
 
 ### Patch Changes