+++ packages/tests/pkg-tests-specs-fixtures/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @business-as-code/tests-specs-fixtures
 
+## 0.0.0-bollards-20230521111204
+
+### Patch Changes
+
+- snapshot
+
 ## 0.0.0-bollards-20230519070635
 
 ### Patch Changes